client: feat(shared/infrastructure/network.service): move NetworkService from core to shared folder
